Fixed Issue #44 (Caption size should be stored as a double).

+++ trunk/VisiGraph/src/edu/belmont/mth/visigraph/models/Caption.java
@@ -16,7 +16,7 @@ public class Caption extends ObservableBase
 	public final Property<Double> x;
 	public final Property<Double> y;
 	public final Property<String> text;
-	public final Property<Integer> size;
+	public final Property<Double> size;
 	public final Property<Boolean> isSelected;
 	
 	public Caption(double x, double y)
@@ -29,17 +29,17 @@ public class Caption extends ObservableBase
 		this(x, y, text, UserSettings.instance.defaultCaptionFontSize.get());
 	}
 	
-	public Caption(double x, double y, String text, int size)
+	public Caption(double x, double y, String text, double size)
 	{
 		this(x, y, text, size, UserSettings.instance.defaultCaptionIsSelected.get());
 	}
 	
-	public Caption(double x, double y, String text, int size, boolean isSelected)
+	public Caption(double x, double y, String text, double size, boolean isSelected)
 	{
 		this.x = new Property<Double>(x, "x");
 		this.y = new Property<Double>(y, "y");
 		this.text = new Property<String>(text, "text");
-		this.size = new Property<Integer>(size, "size");
+		this.size = new Property<Double>(size, "size");
 		this.isSelected = new Property<Boolean>(isSelected, "isSelected");
 	}
 
@@ -48,7 +48,7 @@ public class Caption extends ObservableBase
 		this.x = new Property<Double>((Double)members.get("x"), "x");
 		this.y = new Property<Double>((Double)members.get("y"), "y");
 		this.text = new Property<String>((String)members.get("text"), "text");
-		this.size = new Property<Integer>((Integer)members.get("size"), "size");
+		this.size = new Property<Double>((Double)members.get("size"), "size");
 		this.isSelected = new Property<Boolean>((Boolean)members.get("isSelected"), "isSelected");
 	}
